Sex up templates with static-mockup design
Modified/klooged the app to fallback and serve any additional requested
content from the resources/ folder. Failing to do this causes the
handler to attempt to return everything as text/html.

This has a fair amount of duplication with static-mockup, but until we
decide on a design we'll have to live with this.